Extract UIRefreshControl test helper to a separate file
