+++ EssentialFeed/EssentialFeediOSTests/FeedViewControllerTests.swift
@@ -459,18 +459,6 @@ extension FeedViewController {
         refreshControl = fake
     }
 }
-
-private extension UIRefreshControl {
-    func simulatePullToRefresh() {
-        allTargets.forEach { target in
-            actions(
-                forTarget: target,
-                forControlEvent: .valueChanged)?
-                .forEach({ (target as NSObject).perform(Selector($0))})
-        }
-    }
-}
-
 private class FakeRefreshcontrol: UIRefreshControl {
     private var _isRefreshing = false
