+++ EssentialFeed/EssentialFeediOSTests/Helpers/UIRefreshControl+TestHelpers.swift
@@ -1,0 +1,20 @@
+//
+//  UIRefreshControl+TestHelpers.swift
+//  EssentialFeediOSTests
+//
+//  Created by Marija Zdravic on 14.07.2025..
+//
+
+import Foundation
+import UIKit
+
+extension UIRefreshControl {
+    func simulatePullToRefresh() {
+        allTargets.forEach { target in
+            actions(
+                forTarget: target,
+                forControlEvent: .valueChanged)?
+                .forEach({ (target as NSObject).perform(Selector($0))})
+        }
+    }
+}
